Support for x86_64-w64-mingw32.
On x86-64-MinGW the size of long is 4.  As long is used for
SCM_FIXNUM_BIT, that would mean incompatible .go files, and waste of
cell space.  So we would like to use long long, but the GMP interface
uses long.

To get around this, the x86-64-MinGW port now requires the use of
mini-GMP.  Mini-GMP has been changed to use intptr_t and uintptr_t.

Likewise, "integers.{h,c}" and "numbers.{h,c}" now use intptr_t instead
of scm_t_inum or long, and uintptr_t instead of unsigned long.

* configure.ac: When x86_64-w64-mingw32, require mini-GMP.
* libguile/array-map.c (array_compare): Use uintptr_t instead of
  unsigned long and intptr_t instead of long.
* libguile/arrays.c (make-shared-array): Use ssize_t instead of long
  (scm_array_in_bound_p, scm_array_contents): prefer intptr_t to long
* libguile/mini-gmp.h: Use intptr_t instead of long, uintptr_t instead
    of unsigned long throughout.
* libguile/mini-gmp.c: Likewise.
* libguile/scm.h (SCM_INTPTR_T_BIT): New define.
* libguile/numbers.h (SCM_FIXNUM_BIT): Use it.
* libguile/numbers.c (L1, UL1): New macros.  Use them thoughout instead
  of 1L, 1UL.
  (verify): Use SCM_INTPTR_T_BIT.
  (verify): Use SCM_INTPTR_T_MAX and SCM_INTPTR_T_MIN.
  (scm_from_inum): Remove macro.
  Use intptr_t and uintptr_t instead of scm_t_inum or long, and unsigned
  long.
* libguile/numbers.h (scm_from_intptr, scm_from_uintptr, scm_to_intptr,
    scm_to_uintptr): New defines.
* libguile/integers.h: Use intptr_t and uintptr_t instead of scm_t_inum
    and unsigned long.
* libguile/integers.c (L1) : New macro.  Use it thoughout instead of 1L.
    Use intptr_t and uintptr_t instead of long and unsigned long.
    (long_magnitude): Rename to...
    (intptr_t_magnitude): ...this.  Use intptr_t, uintptr_t.
    (negative_long): Rename to...
    (negative_t_intptr): ...this.  Use uintptr_t, INTPTR_MIN.
    (inum_magnitude): Use intptr_t.
    (ulong_to_bignum): Rename to...
    (uintptr_t_to_bignum): ...this.  Use uintptr_t.
    (long_to_bignum): Rename to...
    (intptr_t_to_bignum): ...this.  Use intptr_t.
    (long_to_scm): Rename to...
    (intptr_t_to_scm): ...this.  Use intptr_to_bignum.
    (ulong_to_scm): Rename to...
    (uintptr_t_to_scm): ...this.  Use uintptr_to_bignum.
    (long_sign): Rename to..
    (intptr_t_sign): ...this.  Use SCM_SIZEOF_INTPTR_T.
    (bignum_cmp_long): Rename to...
    (bignum_cmp_intptr_t): ...this.  Use uintptr_t.
* libguile/bytevectors.c (is_signed_int32, is_unsigned_int32)
    [MINGW32 && __x86_64__]: Use ULL.
    (twos_complement): Use uintptr_t instead of unsigned long.
* libguile/hash.c (JENKINS_LOOKUP3_HASHWORD2): Likewise.
    (narrow_string_hash, wide_string_hash, scm_i_string_hash,
    scm_i_locale_string_hash, scm_i_latin1_string_hash,
    scm_i_utf8_string_hash, scm_i_struct_hash, scm_raw_ihashq,
    scm_raw_ihash): Use and return uintptr_t instead of unsigned long.
    (scm_hashv, scm_hash): Use SCM_UINTPTR_T_MAX.
* libguile/hash.h (scm_i_locale_string_hash, scm_i_latin1_string_hash,
    scm_i_utf8_string_hash): update prototypes.
* libguile/scmsigs.c (sigaction): Use intptr_t instead of long.
* libguile/strings.c (scm_i_make_symbol, (scm_i_c_make_symbol): Use
    uintptr_t instead of unsigned long.
* libguile/strings.h (scm_i_make_symbol, (scm_i_c_make_symbol): Update
    declacations.
* libguile/srfi-60.c: Use scm_to_uintptr, scm_from_intptr and variants
    throughout.
* libguile/symbols.c (symbol-hash): Use scm_from_uintptr.

    Authored-by: Jan (janneke) Nieuwenhuizen <janneke@gnu.org>

+++ libguile/strings.c
@@ -1,4 +1,4 @@
-/* Copyright 1995-1996,1998,2000-2001,2004,2006,2008-2016,2018-2019
+/* Copyright 1995-1996,1998,2000-2001,2004,2006,2008-2016,2018-2019,2021
      Free Software Foundation, Inc.
 
    This file is part of Guile.
@@ -1219,7 +1219,7 @@ SCM_DEFINE (scm_string_ref, "string-ref", 2, 0, 0,
 #define FUNC_NAME s_scm_string_ref
 {
   size_t len;
-  unsigned long idx;
+  uintptr_t idx;
 
   SCM_VALIDATE_STRING (1, str);
 
@@ -1256,7 +1256,7 @@ SCM_DEFINE (scm_string_set_x, "string-set!", 3, 0, 0,
 #define FUNC_NAME s_scm_string_set_x
 {
   size_t len;
-  unsigned long idx;
+  uintptr_t idx;
 
   SCM_VALIDATE_STRING (1, str);
